+++ src/Components/Charts/Donut.php
@@ -198,7 +198,7 @@ class Donut extends Component
 
         for ($i = 0; $i < 3; $i++) {
             $dec = hexdec(substr($hex, $i * 2, 2));
-            $dec = min(max(0, $dec + $dec * $percent), 255);
+            $dec = (int) min(max(0, $dec + $dec * $percent), 255);
 
             $new .= str_pad(dechex($dec), 2, '0', STR_PAD_LEFT);
         }
